refactor(frontend): extract shared AssistantMessageContent component
ChatContent.tsx and onboarding/page.tsx each carried duplicated copies
of normalizeBlockquotes, parseAllBlocks (opportunity / intent_proposal /
networks_panel fenced-block parser), the MessageSegment type, and the
assistant markdown + card rendering. Extract all of it into
chat/AssistantMessageContent.tsx and import from both call sites
(~500 lines deduplicated).

Also adds a 'Continue' suggestion chip to the onboarding communities
step for skipping network joins.
