+++ frontend/src/app/onboarding/page.tsx
@@ -11,15 +11,8 @@ import { useNetworksState } from "@/contexts/IndexesContext";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 import { apiClient } from "@/lib/api";
-import OpportunityCard, {
-  type OpportunityCardData,
-  OpportunitySkeleton,
-} from "@/components/chat/OpportunityCardInChat";
-import IntentProposalCard, {
-  type IntentProposalData,
-  IntentProposalSkeleton,
-} from "@/components/chat/IntentProposalCard";
 import { ToolCallsDisplay } from "@/components/chat/ToolCallsDisplay";
+import AssistantMessageContent from "@/components/chat/AssistantMessageContent";
 import { SuggestionChips } from "@/components/chat/SuggestionChips";
 import { MentionsTextInput } from "@/components/MentionsInput";
 import { cn } from "@/lib/utils";
@@ -47,6 +40,9 @@ const ONBOARDING_STEP_SUGGESTIONS: Record<string, Suggestion[]> = {
   gmail: [
     { label: "Skip for now", type: "direct", followupText: "Skip for now" },
   ],
+  communities: [
+    { label: "Continue", type: "direct", followupText: "I'll skip joining networks for now, let's continue" },
+  ],
   intent: [
     { label: "Building something", type: "prompt", prefill: "Building something " },
     { label: "Exploring partnerships", type: "prompt", prefill: "Exploring partnerships " },
@@ -65,239 +61,6 @@ function buildGreeting(hasName: boolean, userName?: string): string {
   return hasName
     ? `${GREETING_PREAMBLE}\nYou're ${userName}, right? Is that right?`
     : `${GREETING_PREAMBLE} What's your name, and what's your LinkedIn, Twitter/X, or GitHub?`;
-}
-
-// ---------------------------------------------------------------------------
-// Markdown / block parsing (mirrors ChatContent logic)
-// ---------------------------------------------------------------------------
-
-function normalizeBlockquotes(text: string): string {
-  let out = text.replace(/^(>.*?\.\.\.)\s*(\S.+)$/gm, "$1\n\n$2");
-  out = out.replace(/^(>.*)\n(?!>|\n)/gm, "$1\n\n");
-  return out;
-}
-
-type MessageSegment =
-  | { type: "text"; content: string }
-  | { type: "opportunity"; data: OpportunityCardData }
-  | { type: "opportunity_loading" }
-  | { type: "intent_proposal"; data: IntentProposalData }
-  | { type: "intent_proposal_loading" }
-  | { type: "networks_panel"; orderedNetworkIds?: string[] }
-  | { type: "networks_panel_loading" };
-
-function parseAllBlocks(content: string): MessageSegment[] {
-  const segments: MessageSegment[] = [];
-  const regex = /```(opportunity|intent_proposal|networks_panel)\s*\n([\s\S]*?)\n```/g;
-  let lastIndex = 0;
-  let match;
-
-  while ((match = regex.exec(content)) !== null) {
-    if (match.index > lastIndex) {
-      const textBefore = content.slice(lastIndex, match.index);
-      if (textBefore.trim()) segments.push({ type: "text", content: textBefore });
-    }
-    const blockType = match[1];
-
-    {
-      if (blockType === "networks_panel") {
-        // Parse independently so an empty or malformed body still renders the panel
-        // (matches ChatContent.tsx behaviour — graceful degradation, not text fallback).
-        try {
-          const bodyStr = match[2].trim();
-          const body = bodyStr ? (JSON.parse(bodyStr) as Record<string, unknown>) : {};
-          const orderedNetworkIds =
-            Array.isArray(body.orderedNetworkIds) &&
-            (body.orderedNetworkIds as unknown[]).every((id) => typeof id === "string")
-              ? (body.orderedNetworkIds as string[])
-              : undefined;
-          segments.push({ type: "networks_panel", orderedNetworkIds });
-        } catch {
-          segments.push({ type: "networks_panel" });
-        }
-      } else {
-        try {
-          const data = JSON.parse(match[2].trim());
-          if (blockType === "opportunity" && data.opportunityId && data.userId) {
-            segments.push({ type: "opportunity", data: data as OpportunityCardData });
-          } else if (blockType === "intent_proposal" && data.proposalId) {
-            segments.push({ type: "intent_proposal", data: data as IntentProposalData });
-          } else {
-            segments.push({ type: "text", content: match[0] });
-          }
-        } catch {
-          segments.push({ type: "text", content: match[0] });
-        }
-      }
-    }
-    lastIndex = match.index + match[0].length;
-  }
-
-  const remaining = content.slice(lastIndex);
-  const partialOpp = remaining.match(/```opportunity/);
-  const partialIntent = remaining.match(/```intent_proposal/);
-  const partialNetworks = remaining.match(/```networks_panel/);
-
-  const candidates = ([partialOpp, partialIntent, partialNetworks] as (RegExpMatchArray | null)[]).filter(
-    (c): c is RegExpMatchArray => c !== null,
-  );
-  const partialMatch = candidates.length > 0
-    ? candidates.reduce((earliest, c) => c.index! < earliest.index! ? c : earliest)
-    : null;
-
-  if (partialMatch) {
-    const textBefore = remaining.slice(0, partialMatch.index!);
-    if (textBefore.trim()) segments.push({ type: "text", content: textBefore });
-    if (partialMatch === partialOpp) {
-      segments.push({ type: "opportunity_loading" });
-    } else if (partialMatch === partialNetworks) {
-      segments.push({ type: "networks_panel_loading" });
-    } else {
-      segments.push({ type: "intent_proposal_loading" });
-    }
-  } else if (remaining.trim()) {
-    segments.push({ type: "text", content: remaining });
-  }
-
-  if (segments.length === 0 && content.trim()) {
-    segments.push({ type: "text", content });
-  }
-  return segments;
-}
-
-function dedupeSegments(segments: MessageSegment[]): MessageSegment[] {
-  const seenOpps = new Set<string>();
-  const seenProposals = new Set<string>();
-  return segments.filter((seg) => {
-    if (seg.type === "opportunity") {
-      if (seenOpps.has(seg.data.opportunityId)) return false;
-      seenOpps.add(seg.data.opportunityId);
-    }
-    if (seg.type === "intent_proposal") {
-      if (seenProposals.has(seg.data.proposalId)) return false;
-      seenProposals.add(seg.data.proposalId);
-    }
-    return true;
-  });
-}
-
-// ---------------------------------------------------------------------------
-// AssistantMessage (simplified for onboarding — no file upload, no mentions)
-// ---------------------------------------------------------------------------
-
-function AssistantMessageContent({
-  content,
-  isStreaming,
-  onOpportunityPrimaryAction,
-  onOpportunitySecondaryAction,
-  opportunityLoadingMap,
-  currentStatusMap,
-  onIntentProposalApprove,
-  onIntentProposalReject,
-  onIntentProposalUndo,
-  intentProposalStatusMap,
-  OAuthLink,
-  onNetworkJoin,
-  pendingNetworkJoinIds,
-}: {
-  content: string;
-  isStreaming: boolean;
-  onOpportunityPrimaryAction?: (id: string, userId: string, role?: string, name?: string) => void;
-  onOpportunitySecondaryAction?: (id: string, userId: string, role?: string, name?: string) => void;
-  opportunityLoadingMap?: Record<string, boolean>;
-  currentStatusMap?: Record<string, string>;
-  onIntentProposalApprove?: (proposalId: string, description: string, networkId?: string) => void;
-  onIntentProposalReject?: (proposalId: string) => void;
-  onIntentProposalUndo?: (proposalId: string) => void;
-  intentProposalStatusMap?: Record<string, "pending" | "created" | "rejected">;
-  OAuthLink?: React.ComponentType<React.ComponentPropsWithoutRef<"a">>;
-  onNetworkJoin?: (networkId: string, networkTitle: string) => void;
-  pendingNetworkJoinIds?: Set<string>;
-}) {
-  const displayed = normalizeBlockquotes(mentionsToMarkdownLinks(content));
-  const showCursor = isStreaming;
-
-  if (!displayed && isStreaming) {
-    return <span className="inline-block w-2 h-4 bg-current animate-pulse" />;
-  }
-
-  const segments = dedupeSegments(parseAllBlocks(displayed));
-
-  return (
-    <div>
-      {segments.map((seg, idx) => {
-        if (seg.type === "text") {
-          const isLast = idx === segments.length - 1;
-          return (
-            <div
-              key={`text-${idx}`}
-              className={cn(
-                "chat-markdown max-w-none",
-                isStreaming && "chat-markdown-streaming",
-                showCursor && isLast && "chat-markdown-typing",
-              )}
-            >
-              <ReactMarkdown
-                remarkPlugins={[remarkGfm]}
-                components={OAuthLink ? { a: OAuthLink } : undefined}
-              >
-                {seg.content}
-              </ReactMarkdown>
-            </div>
-          );
-        }
-        if (seg.type === "opportunity") {
-          return (
-            <div key={seg.data.opportunityId} className="my-3">
-              <OpportunityCard
-                card={seg.data}
-                onPrimaryAction={onOpportunityPrimaryAction}
-                onSecondaryAction={onOpportunitySecondaryAction}
-                isLoading={opportunityLoadingMap?.[seg.data.opportunityId] ?? false}
-                currentStatus={currentStatusMap?.[seg.data.opportunityId]}
-              />
-            </div>
-          );
-        }
-        if (seg.type === "opportunity_loading") {
-          return <div key={`opp-load-${idx}`} className="my-3"><OpportunitySkeleton /></div>;
-        }
-        if (seg.type === "intent_proposal") {
-          return (
-            <div key={seg.data.proposalId} className="my-3">
-              <IntentProposalCard
-                card={seg.data}
-                onApprove={onIntentProposalApprove}
-                onReject={onIntentProposalReject}
-                onUndo={onIntentProposalUndo}
-                currentStatus={intentProposalStatusMap?.[seg.data.proposalId]}
-              />
-            </div>
-          );
-        }
-        if (seg.type === "networks_panel") {
-          return (
-            <div key={`networks-panel-${idx}`} className="my-3">
-              <NetworksPanel
-                onJoin={onNetworkJoin ?? (() => {})}
-                pendingJoinIds={pendingNetworkJoinIds}
-                orderedNetworkIds={seg.orderedNetworkIds}
-              />
-            </div>
-          );
-        }
-        if (seg.type === "networks_panel_loading") {
-          return (
-            <div key={`networks-panel-loading-${idx}`} className="my-3 flex justify-center py-6">
-              <Loader2 className="h-5 w-5 animate-spin text-gray-300" />
-            </div>
-          );
-        }
-        // intent_proposal_loading
-        return <div key={`intent-load-${idx}`} className="my-3"><IntentProposalSkeleton /></div>;
-      })}
-    </div>
-  );
 }
 
 // ---------------------------------------------------------------------------
@@ -343,7 +106,7 @@ export default function OnboardingPage() {
   const [opportunityStatusMap, setOpportunityStatusMap] = useState<Record<string, string>>({});
   const [intentProposalStatusMap, setIntentProposalStatusMap] = useState<Record<string, "pending" | "created" | "rejected">>({});
   const [proposalIntentMap, setProposalIntentMap] = useState<Record<string, string>>({});
-  const [pendingNetworkJoinIds, setPendingNetworkJoinIds] = useState<Set<string>>(new Set());
+  const [networkPanelPendingJoinIds, setNetworkPanelPendingJoinIds] = useState<Set<string>>(new Set());
 
   const hasName = !!user?.name?.trim();
   const fullGreeting = buildGreeting(hasName, hasName ? `**${user?.name?.trim()}**` : undefined);
@@ -411,6 +174,16 @@ export default function OnboardingPage() {
 
 
 
+  // After stream completes with pending network join IDs, clear them and refresh sidebar
+  const prevNetworkJoinLoadingRef = useRef(isLoading);
+  useEffect(() => {
+    if (prevNetworkJoinLoadingRef.current && !isLoading && networkPanelPendingJoinIds.size > 0) {
+      setNetworkPanelPendingJoinIds(new Set());
+      void refreshIndexes();
+    }
+    prevNetworkJoinLoadingRef.current = isLoading;
+  }, [isLoading, networkPanelPendingJoinIds.size, refreshIndexes]);
+
   // Slide transition: sidebar slides in from left, content shifts right
   const [isTransitioning, setIsTransitioning] = useState(false);
   const hasTriggeredRedirect = useRef(false);
@@ -418,6 +191,9 @@ export default function OnboardingPage() {
   useEffect(() => {
     if (!user?.onboarding?.completedAt || hasTriggeredRedirect.current) return;
     hasTriggeredRedirect.current = true;
+
+    // Always refresh so any new memberships appear in the sidebar immediately
+    void refreshIndexes();
 
     // Accept pending invitation deferred from /l/:code (only after onboarding completes)
     const pendingCode = localStorage.getItem('pendingInviteCode');
@@ -536,7 +312,7 @@ export default function OnboardingPage() {
 
   const handleNetworkJoin = useCallback(
     (networkId: string, networkTitle: string) => {
-      setPendingNetworkJoinIds((prev) => new Set([...prev, networkId]));
+      setNetworkPanelPendingJoinIds((prev) => new Set([...prev, networkId]));
       sendOnboardingMessage(`I'd like to join ${networkTitle}`);
     },
     [sendOnboardingMessage],
@@ -551,6 +327,7 @@ export default function OnboardingPage() {
     if (userCount === 0) return hasName ? "identity" : "identity_no_name";
     if (content.includes("does that sound right") || content.includes("here's what i found")) return "profile";
     if (content.includes("connect your google account") || content.includes("connect gmail")) return "gmail";
+    if (content.includes("communities you might find relevant")) return "communities";
     if (content.includes("what are you open to") || content.includes("open to right now")) return "intent";
     return "identity";
   }, [allMessages, chatMessages, hasName]);
@@ -672,7 +449,7 @@ export default function OnboardingPage() {
                           intentProposalStatusMap={intentProposalStatusMap}
                           OAuthLink={OAuthLink}
                           onNetworkJoin={handleNetworkJoin}
-                          pendingNetworkJoinIds={pendingNetworkJoinIds}
+                          networkPanelPendingJoinIds={networkPanelPendingJoinIds}
                         />
                       </article>
                     </div>
